% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrast.R
\name{contrast}
\alias{contrast}
\title{Compute the treatment effects}
\usage{
contrast(sele, coeVec, varMatrix)
}
\arguments{
\item{sele}{a list specifying the positions of
interactions terms that will be used to compute the treatment effects}

\item{coeVec}{the estimates of fixed effects of the linear mixed-effects model}

\item{varMatrix}{the variance matrix of the estimates of the fixed effects}
}
\description{
Compute the treatment effects
}
