% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/danovir.R
\name{danovir}
\alias{danovir}
\title{Denoised analysis of vaccine-induced immune response}
\usage{
danovir(data, type, covariate, reproduce = FALSE, ...)
}
\arguments{
\item{data}{data set that include all the relevant variables}

\item{type}{a vector of length 4 for which the first component specify
if \code{antigen},\code{antibody} interaction should be include in the model,
the second component specifies whether three factor interaction
 \code{treatment-antigen-antibody} should be included in the model,
 the third component specifies whether \code{antigen} should be included
  in the random effects,  and the fourth component specify whether
   \code{antibody} should be included in the random effects.}

\item{covariate}{variables that need to be included in the model}

\item{reproduce}{If true, then model include a random intercept and
variance of error term is feature wise constant. If false, then model include
random intercept, antigen effect and antibody effect. Variance of errow term is constant.}
}
\value{
A \code{danovir} object which is a list of length 3. The first
component  gives the fitting results of the mixed-effects model;
the second component gives the model-based estimates of treatment effects,
while the third component gives the results from traditional approach.
}
\description{
This is the main function of the package. This function outputs the
  denoised estimate of the treatment effects and their variances. The traditional estimates and their variances are also provided which
  can be used to determine how the model based method affects the estimates.
}
\details{
The underlying basic model in the package is


\code{response_{ijk}=mu+alpha_j+beta_k+lambda_l+gamma_{jl}+gamma_{kl}+u_i+e_{ijk}}


Here the meanings of the terms are, respectively,


 - \code{mu}, the intercept


  - \code{alpha_j} , the main antigen effects


  - \code{beta_k} , the main antibody effects


  - \code{lambda_l} , the main treatment effects


  - \code{gamma_{jl}} , the interaction between treatment and antigen


  - \code{gamma_{kl}} , the interaction between treatment and antibody


 - \code{u_i} , the random intercept for subject


  - \code{e_{ijk}} , the error term


There are four additional terms that are optional which
should be chosen by users. These choices
 should be made based on the characteristics
  of the data at hand. These four terms are, respectively


 - \code{gamma_{jk}}, interaction between antigen and antibody


 - \code{gamma_{jkl}}, interaction between treatment, antigen and antibody


 -  \code{u_{ij}}, random antigen effects


 -  \code{u_{ik}}, random antibody effects
}
