% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/danovir.R
\name{danovir}
\alias{danovir}
\title{Denoised analysis of vaccine-induced immune response}
\usage{
danovir(data, type, covariate, control, ...)
}
\arguments{
\item{data}{data set that include all the relevant variables}

\item{type}{a vector of length 4 for which the first component specify
if \code{antigen},\code{antibody} interaction should be include in the model,
the second component specifies whether three factor interaction
 \code{treatment-antigen-antibody} should be included in the model,
 the third component specifies whether \code{antigen} should be included
  in the random effects,  and the fourth component specify whether
   \code{antibody} should be included in the random effects.}

\item{covariate}{variables that need to be included in the model}

\item{control}{parameters that will be used in the algorithm}
}
\value{
A list of length 2 in which the first
component gives the estimates of treatment effects from the model
while the second component gives the results from tradtional approach
}
\description{
This is the main function of the package. This function outputs the
  denoised estimate of the treatment effects and their variances. The traditional estimates and their variances are also provided which
  can be used to determine how the model based method affects the estimates.
}
